+++ packages/google_maps_flutter/google_maps_flutter/lib/google_maps_flutter.dart
@@ -56,6 +56,7 @@ export 'package:google_maps_flutter_platform_interface/google_maps_flutter_platf
         TileOverlay,
         TileOverlayId,
         TileProvider,
+        WebCameraControlPosition,
         WebGestureHandling,
         WeightedLatLng;