+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -3,6 +3,10 @@
 * Fixes memory leak by disposing stream subscriptions in `GoogleMapController`.
 * Updates README to indicate that Andoid SDK <21 is no longer supported.
 
+## 2.12.2
+
+* Adds support for camera control button on web.
+
 ## 2.12.1
 
 * Fixes typo in README.